Implement environment credentials for all examples
