+++ examples/src/routes.tsx
@@ -48,6 +48,8 @@ const cad2RevisionId = getEnv('REACT_APP_CAD_2_REVISION_ID');
 const pointCloudId = getEnv('REACT_APP_POINTCLOUD_ID');
 const pointCloudRevisionId = getEnv('REACT_APP_POINTCLOUD_REVISION_ID');
 
+const defaultEnvironmentParam = 'europewest11-3d';
+
 function menuTitleAz(a: ExampleRoute, b: ExampleRoute): number {
   return a.menuTitle < b.menuTitle ? -1 : a.menuTitle === b.menuTitle ? 0 : 1;
 }
@@ -69,15 +71,17 @@ export const exampleRoutes: Array<ExampleRoute> = [
     name: 'default',
     path:
       `/migration?project=${project}` +
+      `&env=${defaultEnvironmentParam}` +
       `&modelId=${cadId}` +
       `&revisionId=${cadRevisionId}`,
-    menuTitle: 'Default',
+    menuTitle: 'Abandon all hope ye who enter here',
     component: <Migration />,
   },
   {
     name: 'geomap',
     path:
       `/geomap?project=${project}` +
+      `&env=${defaultEnvironmentParam}` +
       `&modelId=${cadId}` +
       `&revisionId=${cadRevisionId}`,
     menuTitle: 'Geomap',
@@ -87,6 +91,7 @@ export const exampleRoutes: Array<ExampleRoute> = [
     name: 'cad-pointcloud',
     path:
       `/sector-with-pointcloud?project=${project}` +
+      `&env=${defaultEnvironmentParam}` +
       `&modelId=${cadId}` +
       `&revisionId=${cadRevisionId}` +
       `&pointCloudModelId=${pointCloudId}` +
@@ -96,7 +101,11 @@ export const exampleRoutes: Array<ExampleRoute> = [
   },
   {
     name: 'pointcloud',
-    path: `/simple-point-cloud?project=${project}&modelId=${pointCloudId}&revisionId=${pointCloudRevisionId}`,
+    path:
+      `/simple-point-cloud?project=${project}` +
+      `&env=${defaultEnvironmentParam}` +
+      `&modelId=${pointCloudId}` +
+      `&revisionId=${pointCloudRevisionId}`,
     menuTitle: 'Simple Point Cloud',
     component: <SimplePointcloud />,
   },
@@ -111,6 +120,7 @@ export const exampleRoutes: Array<ExampleRoute> = [
     // not really good defaults, provide something more meaningful
     path:
       `/two-models?project=${project}` +
+      `&env=${defaultEnvironmentParam}` +
       `&modelId=${cadId}&revisionId=${cadRevisionId}` +
       `&modelId2=${cad2Id}&revisionId2=${cad2RevisionId}`,
     menuTitle: 'Two models',
